+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # Dev
 - added in skip rounds for masking and selfcal
+- Basic handling of CASDA measurement sets (preprocessing)
+- Basic handling of environment variables in Options, only supported in WSCleanOptions (no need for others yet)
 
 ## 0.2.4